fix(hc): Make RPC pagination args into RpcModels
Resolves errors on generating method signature models for
IntegrationService.
