+++ src/sentry/services/hybrid_cloud/__init__.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import contextlib
-import dataclasses
 import datetime
 import functools
 import inspect
@@ -9,7 +8,6 @@ import logging
 import threading
 from abc import ABC, abstractmethod
 from typing import (
-    TYPE_CHECKING,
     Any,
     Callable,
     Dict,
@@ -26,25 +24,13 @@ from typing import (
     cast,
 )
 
-import sentry_sdk
-from rest_framework.request import Request
-
-from sentry.utils.cursors import Cursor, CursorResult
-from sentry.utils.pagination_factory import (
-    PaginatorLike,
-    annotate_span_with_pagination_args,
-    get_cursor,
-    get_paginator,
-)
-
-logger = logging.getLogger(__name__)
-
 import pydantic
+import sentry_sdk
 
 from sentry.silo import SiloMode
 
-if TYPE_CHECKING:
-    from sentry.api.base import Endpoint
+logger = logging.getLogger(__name__)
+
 T = TypeVar("T")
 
 ArgumentDict = Mapping[str, Any]
@@ -319,81 +305,6 @@ def silo_mode_delegation(
     the mapping values.
     """
     return cast(ServiceInterface, DelegatedBySiloMode(mapping))
-
-
-@dataclasses.dataclass
-class RpcPaginationArgs:
-    encoded_cursor: str | None = None
-    per_page: int = -1
-
-    @classmethod
-    def from_endpoint_request(cls, e: Endpoint, request: Request) -> RpcPaginationArgs:
-        return RpcPaginationArgs(
-            encoded_cursor=request.GET.get(e.cursor_name), per_page=e.get_per_page(request)
-        )
-
-    def do_hybrid_cloud_pagination(
-        self,
-        *,
-        description: str,
-        paginator_cls: Type[PaginatorLike],
-        order_by: str,
-        queryset: Any,
-        cursor_cls: Type[Cursor] = Cursor,
-        count_hits: bool | None = None,
-    ) -> RpcPaginationResult:
-        cursor = get_cursor(self.encoded_cursor, cursor_cls)
-        with sentry_sdk.start_span(
-            op="hybrid_cloud.paginate.get_result",
-            description=description,
-        ) as span:
-            annotate_span_with_pagination_args(span, self.per_page)
-            paginator = get_paginator(
-                None, paginator_cls, dict(order_by=order_by, queryset=queryset.values("id"))
-            )
-            extra_args: Any = {}
-            if count_hits is not None:
-                extra_args["count_hits"] = count_hits
-
-            return RpcPaginationResult.from_cursor_result(
-                paginator.get_result(limit=self.per_page, cursor=cursor, **extra_args)
-            )
-
-
-@dataclasses.dataclass
-class RpcCursorState:
-    encoded: str = ""
-    has_results: bool | None = None
-
-    @classmethod
-    def from_cursor(cls, cursor: Cursor) -> RpcCursorState:
-        return RpcCursorState(encoded=str(cursor), has_results=cursor.has_results)
-
-    # Rpc Compatibility with Cursor
-    def __str__(self) -> str:
-        return self.encoded
-
-    def __bool__(self) -> bool:
-        return bool(self.has_results)
-
-
-@dataclasses.dataclass
-class RpcPaginationResult:
-    ids: List[int] = dataclasses.field(default_factory=list)
-    hits: int | None = None
-    max_hits: int | None = None
-    next: RpcCursorState = dataclasses.field(default_factory=lambda: RpcCursorState())
-    prev: RpcCursorState = dataclasses.field(default_factory=lambda: RpcCursorState())
-
-    @classmethod
-    def from_cursor_result(cls, cursor_result: CursorResult[Any]) -> RpcPaginationResult:
-        return RpcPaginationResult(
-            ids=[row["id"] for row in cursor_result.results],
-            hits=cursor_result.hits,
-            max_hits=cursor_result.max_hits,
-            next=RpcCursorState.from_cursor(cursor_result.next),
-            prev=RpcCursorState.from_cursor(cursor_result.prev),
-        )
 
 
 def coerce_id_from(m: object | int | None) -> int | None:
